Update ze_paper_escaper_p7.cfg
Changed CoolR mode to 1. The cooldown is never activated in the entwatch (so it always shows [R]CoolR:) because the button is parented to a func_physbox_multiplayer, which is parented to a func_door, which is then parented to the weapon_elite. This could probably be fixed via a stripper changing the parentname, but until then this makes the entwatch look better on the hud.