Update ze_stalkermonolit_v8gofix1_override.cfg
The cooldown should not be changed to 8 on the new version of stalker monolit as the item does not behave like a normal item anymore. The flamethrower has an 8 second fuel tank, like flamethrowers on other maps. This fuel tank can be turned off and on while being used. However, when the fuel tank is completly used, THEN it has 8 seconds to reload the entire fuel tank. However, due to a logical error in the map's entity system, the flamethrower requires one to press e twice when the fuel tank is reloaded for it use the fire, and this double press has no cooldown in between. However, since the entwatch prevents one from pressing buttons while items are on cooldown, this no cooldown time is changed to 8 more seconds one cannot use the flamethrower.