Modified Multiple Files (#108)
* Update ze_stalkermonolit_v8gofix1_override.cfg

The cooldown should not be changed to 8 on the new version of stalker monolit as the item does not behave like a normal item anymore. The flamethrower has an 8 second fuel tank, like flamethrowers on other maps. This fuel tank can be turned off and on while being used. However, when the fuel tank is completly used, THEN it has 8 seconds to reload the entire fuel tank. However, due to a logical error in the map's entity system, the flamethrower requires one to press e twice when the fuel tank is reloaded for it use the fire, and this double press has no cooldown in between. However, since the entwatch prevents one from pressing buttons while items are on cooldown, this no cooldown time is changed to 8 more seconds one cannot use the flamethrower.

* Update ze_v0u0v_a6_csgo1.cfg

Make game_player_equip strip weapons on spawn and not give he grenades (which causes players to get 3 a round, instead of 2

* Create ze_mars_escape_gp1.cfg

Remove a door that can block a trigger on the map, leading to delaying

* Update ze_dark_souls_ptd_v0_4_csgo7.cfg

Make humans and zombies have dedicated cages in PvP round so humans don't get stabbed early in cages before CTs can fight back. PvP round is already ZM sided without this, since when a human is stabbed, they turn into a ZM instead of dying.

* Update ze_dreamin_v2_10.cfg

Make buttons on the map easier to press, so you don't do what I did, spamming e on an item SLIGHTLY too high (flipping normal viewheight is too high for some of the item buttons, so you actually have to look down), causing it to not use and lead to your death.