Update ze_genso_of_last_v3_4_t1.cfg
Mapper is endorsing Rulucy's stripper, so might as well enable the easy settings since we get the harder settings from it too (which blocks overdefending level 4 lasers, but not level 3 for some reason. Who cares about consistancy I guess). Genso EW needs ZM heal to be changed to 2 uses with this easify changes.