Update ze_S_A_M_v1_7.cfg
Prevent ZM Ice from freezing itself just above the level 3 water, then unfreezing after the zm tp disables to be in a spot they shouldn't. Yes, there could be a cleaner fix that doesn't always tp them even if not exploiting, but since this method is easy and should work to stop the exploit, it is good enough for me.