Update ze_dark_souls_ptd_v0_4_csgo7_override.cfg
Don't know why the Black flame max uses was removed. It is one of the few items in the map whose max uses do not change throughout the map. The other ZM items gain extra uses if someone presses the pyromancy button at the end of the map in a previous round, but black flame's max uses do NOT increase with the other ZM items.