Modified Multiple Files (#118)
* Update ze_sandstorm_go_v1_5.cfg

Give server settings priority

* Update ze_touhou_gensokyo_o4.cfg

- Make getting a custom model reset player color (doesnt apply to mapper's custom models)
- Give server settings priority, so it is fair to players trying to shoot the secret in spawn (since it can break after a certain number of uses)
- Remove message about chat sounds, since it was removed when mapper backdoor was removed

* Update ze_bathroom_v2_5.cfg

- Fix a new tp avoidance spot
- Prevent delaying for a win at end of level 2 using sponge item as a bridge

* Update ze_dark_souls_ptd_v0_4_csgo7_override.cfg

Don't know why the Black flame max uses was removed. It is one of the few items in the map whose max uses do not change throughout the map. The other ZM items gain extra uses if someone presses the pyromancy button at the end of the map in a previous round, but black flame's max uses do NOT increase with the other ZM items.

* Update ze_touhou_gensokyo_o4.cfg