Modified Multiple Files (#137)
* Update ze_alien_vs_predator_v6go_fix.cfg

* Update ze_ffvii_cosmo_canyon_v5k_fix.cfg

Apparently OnStatTouch doesnt always trigger if you spawn a trigger on top of a player, but does when you enable a trigger they are already in, so this should fix players sometimes getting stuck in earth still

* Update ze_FFXII_Westersand_v7_z9.cfg

Apparently OnStartTouch doesnt always trigger if you spawn a trigger on top of a player, but does when you enable a trigger they are already in, so this should fix players sometimes getting stuck in earth still

* Update ze_MindTree_v1_6.cfg

Fix a broken spawnpoint

* Create ze_tenkinoko_welkin_b6f.cfg

* Update ze_tenkinoko_welkin_b6f.cfg

Add world to end of iceworld and fireworld item names, since they arent standard ice and fire items.

* Update and rename ze_tenkinoko_welkin_b6f.cfg to ze_tenkinoko_welkin_v1.cfg

A new version of the map just came out, so update it to that version.

* Update ze_star_wars_prequel_v1.cfg

Fix max amount of blasters and light sabers. Additionally turn off hud and chat for them, since 20 items being picked up at the same time would spam chat and overflow the !hud menu. Since you don't really need to know who has these items, might as well turn both the chat and hud off and just keep it for the eban protection.