Update README.md
Hopefully this is all correct, since we use savelevel so little that I am not experienced with it. It appears that the match block can also use the ExactMatches, MinMatches, and MaxMatches in conjunction with multiple math, props, or outputs matchs. However, since I haven't seen any SaveLevel config that used this, it doesn't seem necessary to include.