surprise! (#158)
* Update README.md

Hopefully this is all correct, since we use savelevel so little that I am not experienced with it. It appears that the match block can also use the ExactMatches, MinMatches, and MaxMatches in conjunction with multiple math, props, or outputs matchs. However, since I haven't seen any SaveLevel config that used this, it doesn't seem necessary to include.

* Update README.md

Revise wording a bit.

* Update README.md

Spacing

* Update README.md

Spacing

* Update README.md

The datamap is for all entities with targetnames, not just players. Should keep that clear.

* Update README.md

Add example to restoring a datamap

* Update README.md

I had 2 IQ with a replace all for tabs to spaces. Sorry

* Update ze_castle_bridge_v1_3.cfg

cyan isnt a valid chat color

* Update ze_castle_bridge_v1_3.cfg

trigger support. Who needs clean commit history when they get squashed in the merge anyways :)