cruelty squad player_spawn bugs
this script is the same as the one in the map except the player spawn function removes the logic to loop through players, my guess is it doesn't like it on map spawn, which me being lazy i didn't set it up to only fire after warmup, which i probably should have done. this instead changes the logic to fire the fixing function and return the player_spawn function only when the invalid player is found, not double checking all players for valid userids. copium? yes. but in my defense, this shit works fine offline! that's the point of maptests!!